renamed the file and increase test coverage
